Split out header logic
